requirements.txt
lxml==4.9.3
numpy==1.26.4
pillow==11.3.0
setuptools==80.9.0
wheel==0.45.1
flask>=2.0.0
werkzeug>=2.0.0
